+++ tests/e2e/default/client-side-transitions/create-user-for-contacts.create-user.wdio-spec.js
@@ -1,5 +1,6 @@
 const utils = require('../../../utils');
 const sentinelUtils = require('../../../utils/sentinel');
+const messagesUtils = require('../../../utils/messages');
 const placeFactory = require('../../../factories/cht/contacts/place');
 const userFactory = require('../../../factories/cht/users/users');
 const commonPage = require('../../../page-objects/default/common/common.wdio.page');
@@ -26,8 +27,8 @@ describe('Create user when adding contact', () => {
 
   const supervisorUser = utils.deepFreeze(userFactory.build({
     username: 'supervisor',
-    place: district._id, 
-  }));  
+    place: district._id,
+  })); 
 
   beforeEach(async () => {
     await utils.saveDocs([district]);
@@ -41,6 +42,49 @@ describe('Create user when adding contact', () => {
     await utils.revertDb([/^form:/], true);
     await browser.reloadSession();
     await browser.url('/');
+  });
+
+  it.skip('creates a new user while offline', async () => {
+    await utils.updateSettings(settings, 'sentinel');
+    await loginPage.login(supervisorUser);
+    await browser.throttle('offline');
+
+    // TODO create contact
+
+    await browser.throttle('online');
+    await commonPage.syncWithoutWaitForSuccess();
+
+    await sentinelUtils.waitForSentinel();
+    const chwContactId = 'TODO';
+    const { transitions } = await sentinelUtils.getInfoDoc(chwContactId);
+
+    // Transition successful
+    expect(transitions.create_user_for_contacts.ok).to.be.true;
+    // Original contact updated to COMPLETE
+    const finalChwContact = await utils.getDoc(chwContactId);
+    expect(finalChwContact.user_for_contact).to.deep.equal({
+      add: { status: 'COMPLETE' }
+    });
+
+    // New user created
+    const [newUserSettings, ...additionalUsers] = await utils.getUserSettings({ contactId: chwContactId });
+    expect(additionalUsers).to.be.empty;
+    newUsers.push(newUserSettings.name);
+    const loginLink = await messagesUtils.getTextedLoginLink(newUserSettings);
+
+    // TODO log out
+    // Open the texted link
+    await browser.url(loginLink);
+    await commonPage.waitForPageLoaded();
+    const [cookie] = await browser.getCookies('userCtx');
+    expect(cookie.value).to.include(newUserSettings.name);
+  });
+
+  it.skip('creates a new user while online', async () => {
+    await utils.updateSettings(settings, 'sentinel');
+    await loginPage.login(supervisorUser);
+
+    // create contact
   });
 
   it('Creates a new user when adding a person while adding a place', async () => {
@@ -71,6 +115,15 @@ describe('Create user when adding contact', () => {
     const [newUserSettings, ...additionalUsers] = await utils.getUserSettings({ contactId: chwContactId });
     expect(additionalUsers).to.be.empty;
     newUsers.push(newUserSettings.name);
+    
+    const loginLink = await messagesUtils.getTextedLoginLink(newUserSettings);
+    
+    // Open the texted link
+    await commonPage.logout();
+    await browser.url(loginLink);
+    await commonPage.waitForPageLoaded();
+    const [cookie] = await browser.getCookies('userCtx');
+    expect(cookie.value).to.include(newUserSettings.name);
   });
 
   it('Does not create a new user when the transition is disabled', async () => {
@@ -87,9 +140,9 @@ describe('Create user when adding contact', () => {
     const { transitions } = await sentinelUtils.getInfoDoc(chwContactId);
 
     expect(transitions.create_user_for_contacts).to.be.undefined;
+    
 
-    const [newUserSettings, ...additionalUsers] = await utils.getUserSettings({ contactId: chwContactId });
+    const additionalUsers = await utils.getUserSettings({ contactId: chwContactId });
     expect(additionalUsers).to.be.empty;
-    expect(newUserSettings).to.be.undefined; 
   });
 });
